unix: Deduplicate shared os_util code between Linux and macOS
The Linux and macOS os_util.c backends had an identical block:
timing, signal handling, serial (COM) port control, and the TCP helpers.
Move that shared code into a new src/unix/os_util.c compiled into both
platform builds, leaving only the platform-specific functions behind:

- src/linux/os_util.c: CM108 HID PTT, /dev/serial enumeration, GPIO.
- src/macos/os_util.c: CM108 (unsupported stub) and /dev/cu.* enumeration.

The Makefile gains an OBJS_UNIX list that is added to both OBJS_LIN and
OBJS_MAC. This is a pure refactor; the shared code is moved verbatim with
no behavior change.
